add retry logic to TestKV_Patch_RootToken
